requirements.txt
Flask==2.3.0
requests==2.31.0
pyyaml==6.0
pycryptodome==3.18.0
Werkzeug==2.3.0
